Cost functions are now taking the direction of the velocity into account.
Previously, the cost function took only the absolute value of the
translational velocity into account during optimization.
Now the direction is included like intendend initially:
The specified backward-velocity parameter is now taken into account correctly
for the planned trajectory.
A change from a positive to a negative velocity with the same absolute value
does now change the acceleration, which was ignored before.