Obstacle cost function modified (important: read commit description)
Up to this commit, the cost function for obstacle avoidance took
only the projection orthogonal to the trajectory into account instead
of bounding the actual euclidean distance to the obstalce.
One reason for this particular cost function was to avoid pushing trajectory
points along the trajectory rather than to the side in order
to avoid obstacles. But this function introduced a significant drawback:
The projection of the obstacle distance to the normal vector of the trajectory
could be small for points that are far away.
In that case the costs increase to push those poses away, which is undesired!
The effect is getting worse if the config parameter for the number of
affected poses is increased. With the default parameter of 5 affected poses
it's fine as long as the number of obstacles is small. Otherwise the
trajectory starts to jerk around heavily!
I decided to switch back to the natural cost function that
takes only the euclidean distance into account (without projection).
After some tests the results are fine. Pushing poses along the
trajectory is not really recognizable. The only effect that I've recognized is:
If dt_ref is too high (the resolution is too coarse),
it might happen that the trajectory 'jumps' over/through obstacles, since it's
a valid solution.
I will do more tests to figure out a workaround.
But in my opinion the overall navigation behavior is improved with this update.