The package supports carlike robots (ackermann steering) from now on by specifying a minimum bound on the turning radius.
Just specify a turning radius > 0 in the config. Currently, the output of the planner in carlike mode
is still the translational and rotational velocity component at the center of rotation.
I would really appreciate if someone can try this with a real carlike robot and give me some feedback,
since I don't have such a robot ;-)