Fix TebOptimalPlanner::computeError() to use weights.
Call each edge's ->chi2() function, which calculates the squared error
multiplied by the edge's information matrix, which contains the
weighting parameters. Before, the code called
->getError().squaredNorm(), which completely ignores the weights given
by the user.

The need for this arose when I was trying to get the HCP to prefer plans
which minimize backward travel. Despite setting
weight_kinematics_forward_drive very high, the HCP had no preference for
forward-moving TEBs over backward-moving ones. The raw squared error of
the EdgeKinematicsDiffDrive is very small for backward-moving poses,
causing it to be effectively ignored in the plan selection prior to this
change.